// FCAPROGRAMACION/FrontEnd/FCAPROGFE001/src/app/models/Programacion/TipoCaja.ts
export interface TipoCaja {
  idTipoCaja: number;
  claveDiseno: string;
  descripcion: string;
  estatus: boolean;
  regExporta: string;
  permiteCotDirecta: boolean;
  conSuaje: boolean;
  fraccionArancelaria: string;
}
export class TipoCaja {
  idTipoCaja: number;
  claveDiseno: string;
  descripcion: string;
  estatus: boolean;
  regExporta: string;
  permiteCotDirecta: boolean;
  conSuaje: boolean;
  fraccionArancelaria: string;
}